Placing all files under the Txgh namespace
