--- lib/strava/l10n/github_repo.rb
@@ -1,30 +1,0 @@
-require 'config/key_manager'
-require 'strava/l10n/github_api'
-require 'strava/l10n/transifex_project'
-require 'strava/l10n/tx_config'
-
-module Strava
-  module L10n
-    class GitHubRepo
-      attr_reader :name, :branch, :config
-
-      def initialize(name)
-        @name = name
-        Strava::Config::KeyManager.load_yaml(name,nil)
-        @config = Strava::Config::KeyManager.github_repo_config
-        @branch = @config['branch']
-      end
-
-      def transifex_project
-        @transifex_project ||=
-          Strava::L10n::TransifexProject.new(@config['push_source_to'])
-      end
-
-      def api
-        @api ||= Strava::L10n::GitHubApi.new(
-          config['api_username'], config['api_token']
-        )
-      end
-    end
-  end
-end
